src/rememberizer/pyproject.toml
[project]
name = "mcp-server-rememberizer"
version = "0.1.0"
description = "A MCP server project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
 "httpx>=0.27.2",
 "mcp>=1.0.0",
 "python-dotenv>=1.0.1",
]
[[project.authors]]
name = "Rememberizer"
email = "support@rememberizer.ai"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.389",
]

[project.scripts]
rememberizer = "mcp_server_rememberizer:main"